Small gas savings and slot space savings fix

+++ chains/evm/contracts/test/helpers/OffRampHelper.sol
@@ -7,7 +7,7 @@ import {MessageV1Codec} from "../../libraries/MessageV1Codec.sol";
 import {OffRamp} from "../../offRamp/OffRamp.sol";
 
 contract OffRampHelper is OffRamp {
-  uint64 internal constant DEFAULT_MAX_GAS_BUFFER_TO_UPDATE_STATE = 5000 + 5000 + 2000;
+  uint32 internal constant DEFAULT_MAX_GAS_BUFFER_TO_UPDATE_STATE = 5000 + 5000 + 2000;
 
   constructor(
     OffRamp.StaticConfig memory staticConfig
